started refactoring of the back-end

+++ src/main/java/com/compomics/util/experiment/personalization/UrParameter.java
@@ -13,5 +13,5 @@ public interface UrParameter {
      * 
      * @return the parameter key
      */
-    public long getParameterKey();
+    public int getParameterKey();
 }
